Move roundForDisplay to @doenet/utils
The function is now needed in both the worker and the renderer, so it
belongs in the shared utils package. The worker re-exports it from
@doenet/utils for backwards compatibility.

Also fix two pre-existing type errors in math.ts now visible without
the ts-ignore:
- numberToMathExpression: restructure else-if to allow TS to narrow
  the complex-number branch, removing the unreachable else
- preprocessMathInverseDefinition: annotate newAst as [string, ...Tree[]]
  to satisfy me.fromAst()
