+++ packages/doenetml-worker-javascript/src/utils/math.ts
@@ -1,7 +1,9 @@
-// @ts-ignore
 import me from "math-expressions";
+import type { Tree } from "math-expressions";
 
-import { subsets, vectorOperators } from "@doenet/utils";
+import { subsets, vectorOperators, roundForDisplay } from "@doenet/utils";
+
+export { roundForDisplay };
 
 export var appliedFunctionSymbolsDefault = [
     "abs",
@@ -556,17 +558,14 @@ export function mathStateVariableFromNumberStateVariable({
 export function numberToMathExpression(
     number: number | { re: number; im: number },
 ) {
-    let mathTree;
+    let mathTree: Tree;
     if (typeof number === "number") {
         mathTree = number;
-    } else if (
-        typeof number?.re === "number" &&
-        typeof number?.im === "number"
-    ) {
+    } else {
         if (number.im === 0) {
             mathTree = number.re;
         } else {
-            let imPart;
+            let imPart: Tree;
             if (number.im === 1) {
                 imPart = "i";
             } else if (number.im === -1) {
@@ -580,30 +579,8 @@ export function numberToMathExpression(
                 mathTree = ["+", number.re, imPart];
             }
         }
-    } else {
-        mathTree = number;
     }
     return me.fromAst(mathTree);
-}
-
-export function roundForDisplay({ value, dependencyValues }: any) {
-    let rounded = me.round_numbers_to_precision_plus_decimals(
-        value,
-        dependencyValues.displayDigits,
-        dependencyValues.displayDecimals,
-    );
-
-    if (
-        dependencyValues.displayDigits > 0 &&
-        dependencyValues.displaySmallAsZero > 0
-    ) {
-        rounded = me.set_small_zero(
-            rounded,
-            dependencyValues.displaySmallAsZero,
-        );
-    }
-
-    return rounded;
 }
 
 export function mergeListsWithOtherContainers(tree: any) {
@@ -974,7 +951,7 @@ export async function preprocessMathInverseDefinition({
     listIndex = [],
 }: any) {
     if (desiredValue.tree[0] === "list") {
-        let newAst = ["list"];
+        let newAst: [string, ...Tree[]] = ["list"];
         for (let [ind, entry] of desiredValue.tree.slice(1).entries()) {
             let newEntryResult = await preprocessMathInverseDefinition({
                 desiredValue: me.fromAst(entry),
